PROVIDER: Fix recoil bundle issue

+++ src/containers/patients/assets/modal/index.tsx
@@ -1,4 +1,4 @@
-import { memo, Suspense, lazy } from 'react'
+import { memo, Suspense, lazy, useEffect } from 'react'
 import { Modal } from 'components/modal'
 import { Confirm } from 'components/confirm'
 import { useAssetModal } from './use-asset-modal'
@@ -13,33 +13,39 @@ const PatientAssetForm = lazy(() =>
 export const AssetModal = memo(() => {
   const { deleteAsset, dialog, reset } = useAssetModal()
 
-  switch (dialog.type) {
-    case 'asset-edit':
-      return (
-        <Modal
-          aria-labelledby="form"
-          size="md"
-          className="px-10 "
-          onClose={reset}
-          header={dialog.isEditing ? 'Edit asset' : 'Add an new asset'}
-        >
-          <Suspense fallback={<Skeleton />}>
-            <PatientAssetForm />
-          </Suspense>
-        </Modal>
-      )
+  useEffect(() => {
+    return () => reset()
+  }, [])
 
-    case 'asset-delete':
-      return (
-        <Confirm
-          type="delete"
-          description="You are about to delete this asset."
-          onConfirm={deleteAsset}
-          onCancel={reset}
-        />
-      )
+  if (dialog.open)
+    switch (dialog.type) {
+      case 'asset-edit':
+        return (
+          <Modal
+            aria-labelledby="form"
+            size="md"
+            className="px-10 "
+            onClose={reset}
+            header={dialog.isEditing ? 'Edit asset' : 'Add an new asset'}
+          >
+            <Suspense fallback={<Skeleton />}>
+              <PatientAssetForm />
+            </Suspense>
+          </Modal>
+        )
 
-    default:
-      return null
-  }
+      case 'asset-delete':
+        return (
+          <Confirm
+            type="delete"
+            description="You are about to delete this asset."
+            onConfirm={deleteAsset}
+            onCancel={reset}
+          />
+        )
+
+      default:
+        return null
+    }
+  else return null
 })
