+++ src/containers/patients/studies/modal/index.tsx
@@ -1,4 +1,4 @@
-import { memo, Suspense, lazy } from 'react'
+import { memo, Suspense, lazy, useEffect } from 'react'
 import { Modal } from 'components/modal'
 import { Confirm } from 'components/confirm'
 import { usePatientStudiesModal } from './use-patient-studies-modal'
@@ -16,36 +16,42 @@ export const PatientStudiesModal = memo(() => {
 
   const { dialog, reset } = useDialog()
 
-  switch (dialog.type) {
-    case 'patient-study-form':
-      return (
-        <Modal
-          size="sm"
-          className="px-10"
-          onClose={reset}
-          header={
-            dialog.data && dialog.data.isEditing
-              ? 'Edit study'
-              : `Add an study `
-          }
-        >
-          {' '}
-          <Suspense fallback={<Skeleton />}>
-            <PatientStudiesForm />
-          </Suspense>
-        </Modal>
-      )
+  useEffect(() => {
+    return () => reset()
+  }, [])
 
-    case 'patient-study-delete':
-      return (
-        <Confirm
-          type="delete"
-          description={`You are about to delete this study.`}
-          onConfirm={() => mutate()}
-          onCancel={reset}
-        />
-      )
-    default:
-      return null
-  }
+  if (dialog.open)
+    switch (dialog.type) {
+      case 'patient-study-form':
+        return (
+          <Modal
+            size="sm"
+            className="px-10"
+            onClose={reset}
+            header={
+              dialog.data && dialog.data.isEditing
+                ? 'Edit study'
+                : `Add an study `
+            }
+          >
+            {' '}
+            <Suspense fallback={<Skeleton />}>
+              <PatientStudiesForm />
+            </Suspense>
+          </Modal>
+        )
+
+      case 'patient-study-delete':
+        return (
+          <Confirm
+            type="delete"
+            description={`You are about to delete this study.`}
+            onConfirm={() => mutate()}
+            onCancel={reset}
+          />
+        )
+      default:
+        return null
+    }
+  else return null
 })
